CloudSim_Plus_A_Cloud_Computing_Simulati.pdf
Cloud computing research relies on simulation experiments to develop, test, evaluate and tune-up potential solutions. It is difficult to accurately replicate a real system in a simulation experiment, mainly concerned in modelling the arrival of stochastic events such as workload bursts. Therefore, to contribute for valid results, a simulation framework has to 
(i) be well-designed 
(ii) get away from code duplication to avoid code degeneration 
(iii) be extensively tested and 
(iv) provide classes following software engineering principles. 

CloudSimPlus is accordingly aligned with these requirements. It was totally refactored to apply several of these principles to provide a state-of-the-art cloud simulation framework. As future work, there are several features planned to be included, such as SLA management and automatic vertical scaling of Vms.